Split tools out into a seperate module Improve performance test behaviour
