+++ ala-namematching-tools/src/main/resources/log4j.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE log4j:configuration PUBLIC "-//LOGGER" "log4j.dtd">
+
+<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
+
+    <!-- Appenders -->
+    <appender name="console" class="org.apache.log4j.ConsoleAppender">
+        <param name="Threshold" value="info"/>
+        <param name="Target" value="System.out" />
+        <layout class="org.apache.log4j.PatternLayout">
+            <param name="ConversionPattern" value="%d [%t] %-5p: [%c{5}] %m%n" />
+        </layout>
+    </appender>
+
+    <!-- Root Logger -->
+    <root>
+        <level value="warn" />
+        <appender-ref ref="console" />
+    </root>
+
+</log4j:configuration>
