+++ ala-namematching-tools/src/main/resources/standard-suite.csv
@@ -1,0 +1,25 @@
+rps,clients
+100,
+200,
+300,
+400,
+500,
+,1
+,2
+,3
+,4
+,6
+,8
+,10
+,16
+,20,
+,24
+,32
+,40
+,48
+,56
+,64
+,80
+,96
+,128
+,144
