+++ ala-ws-client-testing/src/main/java/au/org/ala/ws/load/LoadTester.java
@@ -4,7 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.io.OutputStreamWriter;
+import java.io.Writer;
+import java.text.DecimalFormat;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -29,6 +30,10 @@ public class LoadTester<T> implements Runnable {
     private LoadSource<T> source;
     /** The clients */
     private List<LoadClient<T>> loads;
+    /** The run start time */
+    private long started;
+    /** The run stop time */
+    private long stopped;
 
     public LoadTester(LoadSource<T> source, int requests, int requestRate, int clients) {
         this.source = source;
@@ -47,6 +52,7 @@ public class LoadTester<T> implements Runnable {
     public void run() {
        List<Thread> runners = this.loads.stream().map(l -> new Thread(l)).collect(Collectors.toList());
        logger.info("Starting run");
+       this.started = System.currentTimeMillis();
        runners.forEach(Thread::start);
        for (Thread runner: runners) {
            try {
@@ -56,16 +62,47 @@ public class LoadTester<T> implements Runnable {
                return;
            }
        }
+       this.stopped = System.currentTimeMillis();
        logger.info("Completed run");
-       Optional<Statistics<Long>> stats = this.loads.stream().map(LoadClient::getStatistics).reduce((s1, s2) -> Statistics.merge(s1, s2));
-       if (stats.isPresent()) {
-           try {
-               stats.get().report(new OutputStreamWriter(System.out));
-           } catch (IOException ex) {
-               logger.error("Unable to write statistics", ex);
-           }
-       } else {
-           logger.warn("No statistics collected");
-       }
+    }
+
+    /**
+     * Create a header for the output
+     *
+     * @param writer The writer to write to
+     *
+     * @throws IOException if unable to writer
+     */
+    public static void reportHeader(Writer writer) throws IOException {
+        writer.write("requests");
+        writer.write(",");
+        writer.write("requestRate");
+        writer.write(",");
+        writer.write("clients");
+        writer.write(",");
+        Statistics.reportHeader(writer);
+        writer.write("\n");
+        writer.flush();
+    }
+
+    public void report(Writer writer) throws IOException {
+        DecimalFormat tFormat = new DecimalFormat("0.00");
+        writer.write(this.requests == 0 ? "" : Integer.toString(this.requests));
+        writer.write(",");
+        writer.write(this.requestRate == 0 ? "" : Integer.toString(this.requestRate));
+        writer.write(",");
+        writer.write(this.clients == 0 ? "" : Integer.toString(this.clients));
+        writer.write(",");
+        Optional<Statistics<Long>> stats = this.loads.stream().map(LoadClient::getStatistics).reduce((s1, s2) -> Statistics.merge(s1, s2));
+        if (stats.isPresent()) {
+            stats.get().report(writer);
+        } else {
+            logger.warn("No statistics collected");
+        }
+        writer.write(",");
+        double rate = this.requests * 1000.0 / Math.max(this.stopped - this.started, 1.0);
+        writer.write(tFormat.format(rate));
+        writer.write("\n");
+        writer.flush();
     }
 }
