+++ pom.xml
@@ -17,6 +17,7 @@
         <module>core</module>
         <module>ala-ws-client-common</module>
         <module>ala-ws-client-testing</module>
+        <module>ala-namematching-tools</module>
     </modules>
 
     <groupId>au.org.ala.names</groupId>
